--- src/frontend/src/components/public/PIAFormTabs/review/pendingReview.tsx
@@ -1,11 +1,0 @@
-const PendingReview = () => {
-  return (
-    <>
-      <div className="review-container ">
-        <div className=" ">Pending review</div>
-      </div>
-    </>
-  );
-};
-
-export default PendingReview;
